+++ release-notes.md
@@ -1,5 +1,65 @@
 # Reply Release Notes
 
+## [v1.0.0](https://github.com/ooaklee/reply/releases/tag/v1.0.0)
+2021-09-11
+
+* Update top-level response members from `meta`, `status` and `data` **->** `meta`, `errors` and `data`
+* Updated `Manifest Error Item` fields
+* Updated logic & added new aide `NewHTTPMultiErrorResponse` to support multiple error response objects in response
+* Refactored code
+* Added logic to support users passing custom *error transfer objects* (`TransferObjectError`), `reply.WithTransferObjectError`
+  * Added `RefreshTransferObject` method call to `TransferObjectError`
+* Updated Token attributes & methods on base `Transfer Object` to have a generic name to cover cases where API uses different token identifiers
+  * `AccessToken` -> `TokenOne` & `RefreshToken` -> `TokenTwo`
+* Added logic to set `StatusCode` if not set already. Defaults to `400`.
+  
+## [v1.0.0-alpha.3](https://github.com/ooaklee/reply/releases/tag/v1.0.0-alpha.3)
+2021-09-11
+
+* Refactor Token attributes & methods to have a more general name to make use of cases where API uses different token identifiers less confusing
+  * `AccessToken` -> `TokenOne` & `RefreshToken` -> `TokenTwo`
+    * Dev can create custom `TransferObject` to set JSON attribute to match their use case
+* Added logic to set `StatusCode` if not set already. Defaults to `400`.
+* Updated README to better describe the library
+
+## [v1.0.0-alpha.2](https://github.com/ooaklee/reply/releases/tag/v1.0.0-alpha.2)
+2021-09-10
+
+* Refactored logic to allow users to pass custom error transfer objects (TransferObjectError), `reply.WithTransferObjectError`
+  * Added `RefreshTransferObject` method call to `TransferObjectError`
+* Updated `example simple api` to use the new Replier Option and update the `transfer object error`
+
+## [v1.0.0-alpha.1](https://github.com/ooaklee/reply/releases/tag/v1.0.0-alpha.1)
+2021-09-10
+
+* Added new aide `NewHTTPMultiErrorResponse` to support multiple error response
+* Updated `example simple api` to use the new aide `NewHTTPMultiErrorResponse`
+* Added logic to handle/ create multiple error response
+* Refactor code to make it more readable with new logic
+
+## [v1.0.0-alpha](https://github.com/ooaklee/reply/releases/tag/v1.0.0-alpha)
+2021-09-09
+
+* Update top-level response members from `meta`, `status` and `data` **->** `meta`, `errors` and `data`
+* Updated underlying logic to how an error is handled
+* Updated `Manifest Error Item` attributes
+
+## [v0.2.0](https://github.com/ooaklee/reply/releases/tag/v0.2.0)
+2021-09-04
+
+* Fixed bug in logic for merging error manifests
+* Added helper functions (aides) to help users more efficiently use the library
+
+## [v0.2.0-alpha.1](https://github.com/ooaklee/reply/releases/tag/v0.2.0-alpha.1)
+2021-09-04
+
+* Fixed bug in logic for merging error manifests
+
+## [v0.2.0-alpha](https://github.com/ooaklee/reply/releases/tag/v0.2.0-alpha)
+2021-09-03
+
+* Initial logic for helper functions (`aides`) to help users more efficiently use the library
+
 ## [v0.1.1](https://github.com/ooaklee/reply/releases/tag/v0.1.1)
 2021-08-31